enlarge scrollable area of the TransactionDetailsViewController
